Handle static functions. Narrow sleds considering reordering
Change name of cfg to function cfg to reflect that the cfg in this
application is different: it uses functions as vertices instead of the
usual basic blocks

The implementation of json printing the cfg has been moved to inside the
cfg class

Target lists now contain function references instead of function names.

Accessing functions in the cfg is different: instead of passing the
function name you need to pass the uniq_label of a function. That is,
    <assembly file of function>.<assembly name of function>
Notice the dot separating the the two fields above.

An 'options' object is passed around to make an options interface easier
to implement

Sleds are narrowed but incorrectly. This will be fixed in the next big
update

Static functions are now handled. Each function is given a unique label
to (as above) so that they can be unambiguously obtained from the cfg